Move debug output system to the context
